Cambio de tipo en cantidad de double a decimal

+++ Clases/MovimientoDeArticulos.cs
@@ -52,7 +52,7 @@ namespace Hamekoz.Negocio
 			set;
 		}
 
-		public double Cantidad {
+		public decimal Cantidad {
 			get;
 			set;
 		}
